strcture refactor
