+++ lib/config/routes/routes.dart
@@ -1,0 +1,6 @@
+final String HOME_ROUTE = '/home';
+final String INTRO_ROUTE = '/intro';
+final String SEND_ROUTE = '/send';
+final String RECEIVE_ROUTE = '/receive';
+final String SPLASH_ROUTE = '/';
+final String SETTINGS_ROUTE = '/settings';
