+++ lib/constants/app_constants.dart
@@ -14,20 +14,12 @@ final String NEXT = 'Next';
 final String CANCEL = 'Cancel';
 final String SELECT_DEFAULT_SAVE_DESTINATION_FOR_THIS_DEVICE = 'Select default save destination for this device.';
 final String CURRENT_SAVE_DESTINATION = 'Current save destination:';
-
-//routes
-final String HOME_ROUTE = '/home';
-final String INTRO_ROUTE = '/intro';
-final String SEND_ROUTE = '/send';
-final String RECEIVE_ROUTE = '/receive';
-final String SPLASH_ROUTE = '/';
-final String SETTINGS_ROUTE = '/settings';
+final String GENERATION_CODE = 'Generating Code';
+final String SEEN= 'seen';
+final String LOADING= 'Loading...';
 
 //keys for tests
-final String HOME_SCAFFOLD = 'HOME_SCAFFOLD';
-final String HOME_MAIN_CONTAINER = 'HOME_MAIN_CONTAINER';
-final String HOME_MAIN_HEADING = 'HOME_MAIN_HEADING';
-
+//-General tests
 final String BOTTOM_NAV_BAR = 'BOTTOM_NAV_BAR';
 final String CUSTOM_NAV_BAR = 'CUSTOM_NAV_BAR';
 
@@ -38,8 +30,9 @@ final String SEND_SCREEN_HEADING = 'SEND_SCREEN_HEADING';
 final String SEND_SCREEN_BOTTOM_SPACE_PLACEHOLDER = 'SEND_SCREEN_BOTTOM_SPACE_PLACEHOLDER';
 final String SEND_SCREEN_CODE_GENERATION_UI = 'SEND_SCREEN_CODE_GENERATION_UI';
 final String SEND_SCREEN_SELECT_A_FILE_BUTTON = 'SEND_SCREEN_SELECT_A_FILE_BUTTON';
+final String GENERATION_DESCRIPTION = 'Generation_Description';
 
-//-receive screen
+//-Receive screen
 final String RECEIVE_SCREEN_BODY = 'RECEIVE_SCREEN_BODY';
 final String RECEIVE_SCREEN_CONTENT = 'RECEIVE_SCREEN_CONTENT';
 final String RECEIVE_SCREEN_HEADING = 'RECEIVE_SCREEN_HEADING';
@@ -49,21 +42,6 @@ final String RECEIVE_SCREEN_ENTER_CODE= 'RECEIVE_SCREEN_ENTER_CODE';
 final String SPLASH_SCREEN_BODY= 'SPLASH_SCREEN_BODY';
 final String SPLASH_SCREEN_LOADING= 'SPLASH_SCREEN_LOADING';
 
-//Strings
-final String SEEN= 'seen';
-final String LOADING= 'Loading...';
 
 
-//fonts
-final String MONTSERRAT = 'Montserrat';
-final String ROBOTO = 'Roboto';
-final String LATO = 'Lato';
 
-//images
-final String SEND_ICON = 'assets/images/send.png';
-final String RECEIVE_ICON = 'assets/images/receive.png';
-final String SETTINGS_ICON = 'assets/images/SETTINGS-WHITE.png';
-final String SETTINGS_ICON_WITH_CIRCLE = 'assets/images/icons/SETTINGS-WHITE-with-circle.png';
-final String LOGO = 'assets/images/logo.png';
-final String PHONE_ICON = 'assets/images/icons/DEVICEWHITE.png';
-final String COPY_ICON = 'assets/images/icons/Paste-white.png';
