+++ lib/constants/asset_path.dart
@@ -1,0 +1,13 @@
+//fonts
+final String MONTSERRAT = 'Montserrat';
+final String ROBOTO = 'Roboto';
+final String LATO = 'Lato';
+
+//images
+final String SEND_ICON = 'assets/images/send.png';
+final String RECEIVE_ICON = 'assets/images/receive.png';
+final String SETTINGS_ICON = 'assets/images/SETTINGS-WHITE.png';
+final String SETTINGS_ICON_WITH_CIRCLE = 'assets/images/icons/SETTINGS-WHITE-with-circle.png';
+final String LOGO = 'assets/images/logo.png';
+final String PHONE_ICON = 'assets/images/icons/DEVICEWHITE.png';
+final String COPY_ICON = 'assets/images/icons/Paste-white.png';
