+++ lib/main.dart
@@ -1,15 +1,12 @@
-import 'dart:io';
-
-import 'package:dart_wormhole_gui/constants/app_constants.dart';
 import 'package:flutter/material.dart';
+import 'config/routes/routes.dart';
 import 'views/introduction-slider.dart';
 import 'views/settings.dart';
 import 'views/splash.dart';
-import 'views/send.dart';
+import 'views/send/send.dart';
 import 'package:flutter/widgets.dart';
-import 'views/receive.dart';
+import 'views/receive/receive.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:window_size/window_size.dart';
 import 'package:desktop_window/desktop_window.dart';
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
