+++ lib/views/receive/receive.dart
@@ -1,8 +1,9 @@
+import 'package:dart_wormhole_gui/config/routes/routes.dart';
 import 'package:dart_wormhole_gui/constants/app_constants.dart';
-import 'package:dart_wormhole_gui/widgets/Button.dart';
-import 'package:dart_wormhole_gui/widgets/EnterCode.dart';
+import 'package:dart_wormhole_gui/widgets/buttons/Button.dart';
+import 'package:dart_wormhole_gui/views/receive/widgets/EnterCode.dart';
 
-import '../widgets/Heading.dart';
+import '../../widgets/Heading.dart';
 import 'package:dart_wormhole_gui/widgets/custom-app-bar.dart';
 import 'package:dart_wormhole_gui/widgets/custom-bottom-bar.dart';
 import 'package:flutter/material.dart';
