+++ lib/views/receive/widgets/EnterCode.dart
@@ -2,8 +2,8 @@ import 'package:dart_wormhole_gui/constants/app_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:dart_wormhole_gui/widgets/FileInfo.dart';
-import 'package:dart_wormhole_gui/widgets/RowGroupButtons.dart';
-import '../widgets/Button.dart';
+import 'package:dart_wormhole_gui/views/send/widgets/RowGroupButtons.dart';
+import '../../../widgets/buttons/Button.dart';
 import 'package:dart_wormhole_gui/widgets/Heading.dart';
 
 class EnterCode extends StatelessWidget {
