+++ lib/views/send/widgets/CodeGeneration.dart
@@ -1,9 +1,10 @@
 import 'package:dart_wormhole_gui/constants/app_constants.dart';
+import 'package:dart_wormhole_gui/constants/asset_path.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:dart_wormhole_gui/widgets/FileInfo.dart';
-import 'package:dart_wormhole_gui/widgets/RowGroupButtons.dart';
-import '../widgets/Button.dart';
+import 'package:dart_wormhole_gui/views/send/widgets/RowGroupButtons.dart';
+import '../../../widgets/buttons/Button.dart';
 import 'package:dart_wormhole_gui/widgets/Heading.dart';
 
 class CodeGeneration extends StatelessWidget {
@@ -20,17 +21,17 @@ class CodeGeneration extends StatelessWidget {
     return  Column(
       children: [
         FileInfo(fileSize, fileName),
-        RowGroupButton('Generating Code', code),
+        RowGroupButton(GENERATION_CODE, code),
         Heading(
             title: SHARE_CODE_WITH_RECIPIENT_AND_WAIT_UNTIL_THE_TRANSFER_IS_COMPLETE,
             textAlign:TextAlign.center,
             marginTop: 16.0.h,
             fontSize: 12.sp,
-            key:Key('Generation_Description'),
+            key:Key(GENERATION_DESCRIPTION),
             fontFamily: MONTSERRAT,
             fontWeight: FontWeight.w300,
         ),
-        Button('Cancel', () {
+        Button(CANCEL, () {
         })
       ],
     );
