+++ lib/views/send/widgets/RowGroupButtons.dart
@@ -1,9 +1,10 @@
 import 'package:dart_wormhole_gui/constants/app_constants.dart';
-import 'package:dart_wormhole_gui/widgets/ButtonLinearGradientWithIcon.dart';
+import 'package:dart_wormhole_gui/constants/asset_path.dart';
+import 'package:dart_wormhole_gui/widgets/buttons/ButtonLinearGradientWithIcon.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:dart_wormhole_gui/widgets/Heading.dart';
-import 'ButtonWithIcon.dart';
+import '../../../widgets/buttons/ButtonWithIcon.dart';
 import 'package:flutter/services.dart';
 
 class RowGroupButton extends StatelessWidget {
