+++ lib/views/send/widgets/SendingProgress.dart
@@ -1,10 +1,11 @@
 import 'package:dart_wormhole_gui/constants/app_constants.dart';
+import 'package:dart_wormhole_gui/constants/asset_path.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:dart_wormhole_gui/widgets/Heading.dart';
 
-import 'Button.dart';
-import 'FileInfo.dart';
+import '../../../widgets/buttons/Button.dart';
+import '../../../widgets/FileInfo.dart';
 
 class SendingProgress extends StatelessWidget {
   final int fileSize;
