+++ lib/views/send/widgets/SentSuccessfully.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:dart_wormhole_gui/widgets/Heading.dart';
 
-import 'Button.dart';
-import 'FileInfo.dart';
+import '../../../widgets/buttons/Button.dart';
+import '../../../widgets/FileInfo.dart';
 
 class SentSuccessfully extends StatelessWidget {
   final int fileSize;
