+++ lib/widgets/custom-app-bar.dart
@@ -1,5 +1,7 @@
 import 'package:app_settings/app_settings.dart';
+import 'package:dart_wormhole_gui/config/routes/routes.dart';
 import 'package:dart_wormhole_gui/constants/app_constants.dart';
+import 'package:dart_wormhole_gui/constants/asset_path.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -70,7 +72,6 @@ class _CustomAppBarState extends State<CustomAppBar> {
            Divider(height: 1.0.h, color: Colors.white),
           ],
         )
-
-          );
+    );
   }
 }
